feat(msp): remove old service view
feat(msp): remove old service view
